Recherche recursive

+++ Triage2.py
@@ -1,13 +1,14 @@
-'''Version : 1.1 '''
+'''Version : 1.2 '''
 
-#a faire : lire plus qu'un dossier de profondeur
+#a faire :
+
+# Synchronisation sur base de donnee sur le web? Pour chercher posters et sous-titres
+
 
 import os
 from os import rename, path
 import stat
 from string import capwords
-
-#ceci est dans recherche-recursive
 
 class Item_to_process:
 
@@ -21,94 +22,6 @@ class Item_to_process:
 
         self._prefixe, self._extension = os.path.splitext(self._path_complet)
 
-        print('on process: ' + self._nom_fichier)
-
-
-    def seasonid(self):
-
-        scanner = 0
-
-        while True:
-
-            try:
-
-                if (self._nom_fichier[scanner] == 's' or self._nom_fichier[scanner] == 'S') and (self._nom_fichier[(scanner + 1)] in str(list(range(0, 10)))) and self._nom_fichier[(scanner - 1)] == '.':
-
-                    season_episode = ''.join(self._nom_fichier[scanner:(scanner + 6)])
-
-                    title = ''.join(self._nom_fichier[0:scanner])
-
-                    print('{} detecte {}{}'.format(''.join(self._nom_fichier), title, season_episode))
-                    break
-
-                scanner += 1
-
-            except IndexError:
-                print('ERROR : aucune saison trouve dans {} (.seasonid1)'.format(''.join(self._nom_fichier)))
-                break
-
-    def seasonid2(self):
-
-        scanner = 0
-
-        season_episode = None
-
-        for lettre in self._nom_fichier:
-
-            if (self._nom_fichier[scanner] == 's' or self._nom_fichier[scanner] == 'S') and (self._nom_fichier[(scanner + 1)] in str(list(range(0, 10)))) and self._nom_fichier[(scanner - 1)] == '.':
-
-                season_episode = ''.join(self._nom_fichier[scanner:(scanner + 6)])
-                print('{} |detecte| {}'.format(''.join(self._nom_fichier), season_episode))
-                break
-
-            scanner += 1
-
-        if season_episode == None:
-
-            print('aucune saison trouve dans {} (.seasonid2)'.format(''.join(self._nom_fichier)))
-
-    def seasonid3(self):
-
-        longeur_nom_fichier = len(self._nom_fichier)
-
-
-        # dossier destination pour les films
-        movies_pathto = ('{}\{}'.format(dossier_films, self._nom_fichier))
-
-        season_episode = None
-
-        for i in range(0,longeur_nom_fichier):
-
-            self._nom_fichier = list(self._nom_fichier)
-
-            if (self._nom_fichier[i] in ['s','S']) and (self._nom_fichier[(i + 1)] in str(list(range(0, 10)))) and \
-                    self._nom_fichier[(i - 1)] == '.':
-
-                series_episode = ''.join(self._nom_fichier[i:(i + 6)])
-
-                series_title = ''.join(self._nom_fichier[0:i])
-
-                [print('b') if n == '.' else False for n in series_title]
-
-
-                series_pathto = ('{}\{}\{}'.format(dossier_series, series_title, self._nom_fichier))
-
-                print('{} |detecte| {} {} (.seasonid3)'.format(''.join(self._nom_fichier),series_title, series_episode))
-
-                #rename(self._path_complet, series_pathto)
-
-            if self._nom_fichier[i] in ['1', '2'] and self._nom_fichier[i + 1] in str(list(range(0, 10))) \
-                  and self._nom_fichier[i + 2] in str(list(range(0, 10))) and self._nom_fichier[i + 3] in str(list(range(0, 10))) \
-                  and (self._nom_fichier[i - 1]) in ['.', '(', '[', '-', ' ']:
-
-                movie_title = ''.join(self._nom_fichier[0:i - 2])
-
-                movie_year = ''.join(self._nom_fichier[i:i + 4])
-
-                print('{} |detecte| {} {} (.seasonid3)'.format(''.join(self._nom_fichier), movie_title, movie_year))
-
-                # Deplacement du film vers le dossier film
-                rename(self._path_complet, movies_pathto)
 
     def classify(self):
 
@@ -116,24 +29,28 @@ class Item_to_process:
 
         range_1_a_10 = str(list(range(0, 10)))
         
-        season_episode = None
+        type_detecte = None
 
         for i, n in enumerate(self._nom_fichier):
 
             # On detecte si c'est une serie televisee en detectant le pattern S0102
             try:
 
-                if (self._nom_fichier[i] in ['s', 'S']) and (self._nom_fichier[(i + 1)] in range_1_a_10) and (self._nom_fichier[(i - 1)] in ['.', ' ']):
+                if (self._nom_fichier[i] in ['s', 'S', 'E', 'e']) and (self._nom_fichier[(i + 1)] in range_1_a_10) and (self._nom_fichier[(i - 1)] in ['.', ' ']):
 
-                    series_episode = ''.join(self._nom_fichier[i:(i + 6)])
+                    type_detecte = 'Serie'
+
+                    if self._nom_fichier[i] in ['E', 'e']:
+
+                        series_episode = ('S01' + self._nom_fichier[i:(i + 3)])
+
+                    else:
+
+                        series_episode = ''.join(self._nom_fichier[i:(i + 6)])
 
                     series_episode = series_episode.upper()
 
-                    print('series_episode: ' + series_episode)
-
                     series_title = ''.join(self._nom_fichier[0:(i-1)])
-
-                    print('series_title: ' + series_title)
 
                     series_title_clean = ''
 
@@ -160,18 +77,22 @@ class Item_to_process:
                     # Si le dossier qui contient la saison n'existe pas, on le cree et finalement on deplace vers celui-ci
                     if os.path.isdir('{}\{}\{}'.format(dossier_series, series_title_clean, seasons_folder)):
 
-                        rename(self._path_complet, series_pathto)
+                        print(self._path_complet + '-->' + series_pathto)
+                        if simulation == False:
+                            rename(self._path_complet, series_pathto)
 
                     else:
 
                         os.makedirs('{}\{}\{}'.format(dossier_series, series_title_clean, seasons_folder))
-                        rename(self._path_complet, series_pathto)
 
-                    print('{} |detecte| {} {} (.seasonid3)'.format(''.join(self._nom_fichier), series_title, series_episode))
+                        print(self._path_complet + ' --> ' + series_pathto)
+                        if simulation == False:
+                            rename(self._path_complet, series_pathto)
+
+
 
             except IndexError:
                 pass
-
 
             # On detecte un film en trouvant une annee
 
@@ -179,6 +100,8 @@ class Item_to_process:
                 if self._nom_fichier[i] in ['1', '2'] and self._nom_fichier[i + 1] in str(list(range(0, 10))) \
                         and self._nom_fichier[i + 2] in range_1_a_10 and self._nom_fichier[i + 3] in range_1_a_10 \
                         and (self._nom_fichier[i - 1]) in ['.', '(', '[', '-', ' ']:
+
+                    type_detecte = "Film"
 
                     movie_title = ''
 
@@ -201,14 +124,21 @@ class Item_to_process:
                     # dossier destination pourn les films
                     movies_pathto = ('{}\{} ({}){}'.format(dossier_films, movie_title, movie_year, self._extension))
 
-                    print('{} |detecte| {} {} (.seasonid3)'.format(''.join(self._nom_fichier), movie_title, movie_year))
-
                     # Deplacement du film vers le dossier film
-                    rename(self._path_complet, movies_pathto)
+                    print(self._path_complet + ' --> ' + movies_pathto)
+                    if simulation == False:
+                        rename(self._path_complet, movies_pathto)
 
             except:
 
                 pass
+
+        if type_detecte == None:
+
+            print('Au purgatoire: ' + self._nom_fichier)
+
+            if simulation == False:
+                os.rename(self._path_complet, purgatoire)
 
     def purge(self):
 
@@ -220,16 +150,22 @@ class Item_to_process:
 
             try:
 
-                os.remove(self._path_complet)
                 print('supression du fichier: ' + self._nom_fichier)
+
+                if simulation == False:
+                    os.remove(self._path_complet)
+
 
             except PermissionError:
 
                 print('erreur de permission, deplacement au purgatoire')
-                os.rename(self._path_complet, purgatoire)
 
+                if simulation == False:
+                    os.rename(self._path_complet, purgatoire)
 
-source = os.getcwd()
+simulation = True
+
+source = 'd:\\downloads\\triage'
 
 dossier_films = '{}\Films'.format(source)
 
@@ -237,7 +173,34 @@ dossier_series = '{}\Series'.format(source)
 
 dossiers_base = ['purgatoire','Films','Series']
 
+
+
 dossiers = os.listdir(source)
+
+root = os.walk(source)
+
+for racine, directories, fichiers in root:
+
+    for fichier in fichiers:
+
+        if racine[0:(len(source)+7)] != os.path.join(source,'Series') and racine[0:(len(source)+6)] != os.path.join(source,'Films'):
+
+            individu = Item_to_process(fichier, racine)
+            individu.purge()
+            individu.classify()
+
+#On fait le menage des dossiers vides
+for dossier in os.listdir(source):
+
+    if dossier not in dossiers_base:
+
+        print('Supression du dossier ' + (os.path.join(source, dossier)))
+
+        if simulation == False:
+            os.removedirs(os.path.join(source, dossier))
+
+
+quit()
 
 #Creation des fichiers de base si necessaire
 for i in dossiers_base:
@@ -246,7 +209,6 @@ for i in dossiers_base:
 
         os.makedirs('{}\{}'.format(source, i))
 
-#iteration sur chaqun des items a la source
 for items in dossiers :
 
     #path de l'item itere
